edge/requirements.txt
numpy>=1.24,<3
opencv-python-headless>=4.8,<5
torch>=2.1,<3
torchvision>=0.16,<1
transformers>=4.38,<5
safetensors>=0.4,<1
PyYAML>=6.0,<7
paho-mqtt>=2.0,<3
boto3>=1.34,<2
requests>=2.31,<3
Pillow>=10.0,<12
